+++ Sources/CoreCollector/backend/CallbackEventDataDispatcher.swift
@@ -1,0 +1,4 @@
+internal protocol CallbackEventDataDispatcher {
+    func add(_ eventData: EventData, completionHandler: @escaping (HttpDispatchResult) -> Void)
+    func addAd(_ adEventData: AdEventData, completionHandler: @escaping (HttpDispatchResult) -> Void)
+}