+++ Tests/CoreCollectorTests/backend/AuthenticatedDispatcherTests.swift
@@ -6,7 +6,7 @@ import Quick
 import Foundation
 
 class AuthenticatedDispatcherTests: QuickSpec {
-    override func spec() {
+    override class func spec() {
         var mockAuthService = MockAuthenticationService()
         var mockInnerDispatcher = MockEventDataDispatcher()
         var mockNotificationCenter = NotificationCenter()
@@ -128,7 +128,7 @@ class AuthenticatedDispatcherTests: QuickSpec {
                 verify(mockInnerDispatcher).resetSourceState()
             }
         }
-        describe("authentication failed") {
+        describe("authentication denied") {
             it("should call innerDispatcher") {
                 // arrange
                 stub(mockInnerDispatcher) { stub in
@@ -136,7 +136,42 @@ class AuthenticatedDispatcherTests: QuickSpec {
                 }
 
                 // act
-                mockNotificationCenter.post(name: .authenticationFailed, object: mockAuthService)
+                mockNotificationCenter.post(name: .authenticationDenied, object: mockAuthService)
+
+                // assert
+                verify(mockInnerDispatcher).disable()
+            }
+            it("should not call add and addAd on innerDispatcher when authentication was denied") {
+                // arrange
+                stub(mockInnerDispatcher) { stub in
+                    when(stub.disable()).thenDoNothing()
+                }
+
+                let eventData = EventData("test-impression")
+                let adEventData = AdEventData()
+
+                // emit authentication success event
+                mockNotificationCenter.post(name: .authenticationSuccess, object: mockAuthService)
+
+                // act
+                mockNotificationCenter.post(name: .authenticationDenied, object: mockAuthService)
+
+                // assert
+                verify(mockInnerDispatcher).disable()
+                dispatcher.addAd(adEventData)
+                dispatcher.add(eventData)
+                verifyNoMoreInteractions(mockInnerDispatcher)
+            }
+        }
+        describe("authentication error") {
+            it("should call innerDispatcher") {
+                // arrange
+                stub(mockInnerDispatcher) { stub in
+                    when(stub.disable()).thenDoNothing()
+                }
+
+                // act
+                mockNotificationCenter.post(name: .authenticationError, object: mockAuthService)
 
                 // assert
                 verify(mockInnerDispatcher).disable()
@@ -154,7 +189,7 @@ class AuthenticatedDispatcherTests: QuickSpec {
                 mockNotificationCenter.post(name: .authenticationSuccess, object: mockAuthService)
 
                 // act
-                mockNotificationCenter.post(name: .authenticationFailed, object: mockAuthService)
+                mockNotificationCenter.post(name: .authenticationError, object: mockAuthService)
 
                 // assert
                 verify(mockInnerDispatcher).disable()