+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - [internal] Release/Create Buttons for AV Example App
+- `BitmovinAnalyticsConfig.offlinePlaybackAnalyticsEnabled` with default value of `false`. When set to `true`, analytics data is collected and persisted for offline playback sessions and re-sent as soon as there is network again
 
 ### Removed
 - removed support for iOS/tvOS 12.0, 13.0